organized the code structure and fixed the data refetch issue on add record and update
